✅   Main Manager 🚧   Sistema de vida 🚧  Pausa
✅  Main Manager
 - Se creo el main manager donde se maneja principalmente el sistema de pausa

🚧  Sistema de vida
 - El sistema de vida se implemento al player, de momento se implemento solamente en numeros todavia no afecta a nada

🚧 Sistema de Pausa
- Se implemento el sistema de pausa dentro del player para que no pueda abrir tiendas en pausa y restringir sus movimientos

👷 Siguientes cosas a trabajar:
 - terminar la pausa principalmente con los proyectiles
 - Avanzar con el sistema de vida
 - Empezar con el sistema de estadisticas para incluirlo en las tiendas

+++ Assets/Enemies/Scripts/EnemyDummie.cs
@@ -12,7 +12,7 @@ public class EnemyDummie : MonoBehaviour, IDamageable
         currentHealth = maxHealth;
     }
  
-    public void TakeDamage(float amount)
+    public void TakeDamage(int amount)
     {
         currentHealth -= amount;
         Debug.Log($"{name} recibió {amount} de daño. Vida restante: {currentHealth}");
