+++ Assets/Interfaces/IDamageable.cs
@@ -2,5 +2,5 @@
 // debe implementar esta interfaz.
 public interface IDamageable
 {
-    void TakeDamage(float amount);
+    void TakeDamage(int amount);
 }
